Remove SET_STAGE_ENABLED and SET_KERNEL_ENABLED
We no longer have a situation in which a driver is built but would not
be enabled. Drivers that have dependencies will simply not be built if
the dependency is not met; this is controlled by CMake. Running `pdal
drivers` will still provide visibility at runtime into which drivers are
available.