Add PCL kernel to compute Height Above Ground
Requires either 1) classification dimension with ground returns in the input or
2) a secondary input that is only ground returns.

Computes Height Above Ground as difference in Z between the input point and the
nearest corresponding ground return (in the XYZ plane).

Reader drivers are inferred from the filenames, but the Writer must be BPF to
accomodate the non-standard dimension.