Add writers.pcd.subtract_minimum
New option writers.pcd.subtract_minimum (bool default true)

Restore previous behavior of defaulting to subtract minimum XYZ
dimension before writing.
The options offset_* are applied secondarily to this minimum XYZ
offset. So if subtract_minimum is true (default), offset_* are relative
to the new origin of the min XYZ. If subtract_minimum is false, offset_*
are relative to the original frame.

The offset subtraction occurs before any scaling.

This flag is not stored with the new pcd file in any way.