Shrink the PCL plugin
As discussed in #1404, PDAL can now do many of the things we had been relying
on PCL to do. This commit removes the PCL versions of filters that had been
duplicated across PDAL and PCL, specifically:

* filters.dartsample
* filters.ground
* filters.height
* filters.radiusoutlier
* filters.statisticaloutlier
* kernels.views
* writers.pclvisualizer

Ground kernel is now a native PDAL kernel (it had been part of the PCL plugin).
It now calls filters.pmf (PDAL) as opposed to filters.ground (PCL), which are
both implementations of the PMF algorithm.

We have made an effort to update documentation (especially in the workshop) to
instruction users in the proper usage of the native PDAL capabilities. Many of
the tutorials still refer to the PCL implementations though, as they serve a
dual purpose of introducing the filtering approach and also covering concepts
of how to write plugins for PDAL.

The HeightAboveGround kernel in the PCL plugin had not been built or maintained
in months and has been removed.

To the extent that filters.ground, filters.statisticaloutlier, and
filters.radiusoutlier were being tested via the PCLBlockFilterTest, filters.pmf
and filters.outlier are being tested now in OldPCLBlockTest.